requirements.txt
aiogram==2.25.1
openai==0.27.8
yoomoney==0.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.6 